Cleanup test helpers from VM
