+++ machines/tests/integration/juju_.py
@@ -3,21 +3,8 @@
 
 import importlib.metadata
 
-import juju.unit
-
 # libjuju version != juju agent version, but the major version should be identical—which is good
 # enough to check for secrets
 _libjuju_version = importlib.metadata.version("juju")
 has_secrets = int(_libjuju_version.split(".")[0]) >= 3
 juju_major_version = int(_libjuju_version.split(".")[0])
-
-
-async def run_action(unit: juju.unit.Unit, action_name, **params):
-    action = await unit.run_action(action_name=action_name, **params)
-    result = await action.wait()
-    # Syntax changed across libjuju major versions
-    if juju_major_version <= 2:
-        assert result.results.get("Code") == "0"
-    else:
-        assert result.results.get("return-code") == 0
-    return result.results
